+++ R/linearModelTraining.R
@@ -9,6 +9,11 @@
 #' \code{clsLayer[i,j]} is a cluster ID of individual \code{i} at layer \code{j}; \code{clsLayer[i,1]} is the first layer that everyone typically belongs to a single cluster.
 #'@param insigThs is a threshold to determine whether a magnitude of a feature coefficient is enough so that the feature is designated as a selected feature.
 #'@param alpha is a significance level to determine whether a magnitude of a feature coefficient is enough so that the feature is designated as a selected feature.
+#'@param messageFlag is a flag. If it is true, the function shows the text regarding the progress of computing.
+#'@param polyDegree is a degree of polynomial function that is used to fit the data.
+#'If it is greater than 1, the polynomial formula is used in \code{lm()} instead of \code{"y=."}.
+#'@param expFlag is an exponential flag to control the formula for data fitting.
+#'If it is true, then the exp() formula is used in \code{lm()} instead of \code{"y=."}.
 #'
 #'@return This function returns \code{models} and \code{DataT}.
 #'
@@ -23,10 +28,10 @@
 #' obj<-linearModelTraining(DataT)
 #'
 #'@importFrom caret trainControl train
-#'@importFrom stats cor lm predict rnorm runif
+#'@importFrom stats cor lm predict rnorm runif as.formula
 #'@export
 #'
-linearModelTraining<-function(DataT,insigThs=1e-8,alpha=0.05)
+linearModelTraining<-function(DataT,insigThs=1e-8,alpha=0.05,messageFlag=FALSE,polyDegree = 1, expFlag= FALSE)
 {
 
   out1<-remappingClusterInx(DataT$clsLayer)
@@ -53,7 +58,10 @@ linearModelTraining<-function(DataT,insigThs=1e-8,alpha=0.05)
       y<-DataT$Y[inxFilterVec,]
       df = data.frame(y, x)
       #===========================st Linear model
-      submodels[[inx2]] <- lm(y ~ ., data = df)
+      if(polyDegree == 1 && expFlag == FALSE)
+        submodels[[inx2]] <- lm(y ~ ., data = df)
+      else
+        submodels[[inx2]] <- lm(formula = makePolyFormula(df,degree = polyDegree, expFlag= expFlag), data = df)
       submodels[[inx2]] $ID<-IDk
       IDk<-IDk +1
       #===========================fn Linear model
@@ -85,11 +93,27 @@ linearModelTraining<-function(DataT,insigThs=1e-8,alpha=0.05)
         residualMat[inxFilterVec,inx] <-sum(submodels[[inx2]]$residuals)
 
       submodels[[inx2]]$clsName<-sprintf("C%g",DataT$clsNameMappingTable[[inx]][[inx2]])
-      print(sprintf("Layer%d,Cls:%s",inx,submodels[[inx2]]$clsName))
+      if(messageFlag == TRUE)
+        message(sprintf("Training -> Layer%d,Cls:%s",inx,submodels[[inx2]]$clsName))
     }
 
     models[[inx]]<-submodels
   }
   DataT$nNodes<-nNodes
   return(list("models"=models,"DataT"=DataT))
+}
+
+makePolyFormula<-function(df,degree=2, expFlag= FALSE)
+{
+  varnames <- paste("X", 1:(length(df)-1), sep="")
+  for(i in seq(length(varnames)))
+  {
+    if(expFlag == FALSE)
+      varnames[i]<-sprintf("%s^%d",varnames[i],degree)
+    else
+      varnames[i]<-sprintf("exp(%s)",varnames[i],degree)
+  }
+
+  polyFormula <- as.formula(paste("y ~ ", paste(varnames, collapse= "+")))
+  return(polyFormula)
 }