+++ NEWS.md
@@ -1,2 +1,6 @@
+# MRReg v0.1.1
+* Add polynomial and exponential fitting and simulation
+* Revise the documents
+
 # MRReg v0.1.0
 * The first release version of MRReg